requirements.txt
aniso8601==3.0.0
click==6.7
dill==0.2.5
Flask==1.0.1
Flask-RESTful==0.3.6
flask-restful-swagger-2==0.35
flask-swagger-ui==3.6.0
future==0.16.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyFunctional==1.1.3
pytz==2018.4
six==1.10.0
tabulate==0.7.7
Werkzeug==0.14.1
